Extract drawing code
We'll be able to use this on the wayland backend.
